Functionality for insertCode option
Added module dependency 'fs'.
Passed by parameter the insertCode option on transform method on line 15.
Added getFileContent method on line 50, for get all file content with the filepath, then on transform[...][...] methods checks if exists inside arguments this parameter and then if it's true inject the code, if not then injects the correspondent directive.